Revert "Fixed implicit conversion warning: __LINE__ is a long (#4397)"
This reverts commit 4ff59b715dc329e3c4a3a9def85e4a841f9916c9.

+++ libtransmission/log.h
@@ -52,7 +52,7 @@ struct tr_log_message
 
     // location in the source code
     std::string_view file;
-    long line;
+    size_t line;
 
     // when the message was generated
     time_t when;
@@ -91,7 +91,7 @@ void tr_logSetLevel(tr_log_level);
 
 void tr_logAddMessage(
     char const* source_file,
-    long source_line,
+    int source_line,
     tr_log_level level,
     std::string_view msg,
     std::string_view module_name = {});
